Moved render functions to utils/render and created utils/common for the rest
