+++ src/components/site-navigation.js
@@ -1,4 +1,4 @@
-import {capitalizeFirstLetter} from '../utils';
+import {capitalizeFirstLetter} from '../utils/common';
 import AbstractComponent from './abstract-component';
 
 const createFilterMarkup = (name, isActive, count) => {
